+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 Changelog
 =========
 
-1.4.pre
+1.4.0
 -----------------
+  * removed old php
   * rename sx.config to sx.Config
   * sx.cookie expires 1 year by default
   * Component cookie is @deprecated